Added Dark and Light Themes,extensions with their Colors,refactored some code,changed ColorStyle and FontStyle in App.Tested by available entries of Calories.

+++ lib/models/calorie_max.dart
@@ -1,5 +1,5 @@
 class CalorieMax {
-  final double? calories;
+  final int? calories;
 
   CalorieMax({required this.calories});
 }
