+++ lib/models/meal_expense.dart
@@ -1,6 +1,6 @@
 class MealExpense {
   final DateTime dateTime;
-  final double? calories;
+  final int? calories;
 
   MealExpense({required this.dateTime, required this.calories});
 }
